renamed InferenceProblem to InverseProblem
