+++ probeye/__init__.py
@@ -1,3 +1,3 @@
 # this is the only place to define the package version; it is currently used by
 # setup.cfg, docs/conf.py and by the method print_probeye_header in subroutines.py
-__version__ = "2.1.4"
+__version__ = "2.1.5"
