+++ probeye/definition/__init__.py
@@ -1,4 +1,5 @@
 # module imports
+from probeye.definition import inverse_problem
 from probeye.definition import inference_problem
 from probeye.definition import forward_model
 from probeye.definition import likelihood_model
