+++ tests/integration_tests/tests_without_correlation/test_maximum_likelihood.py
@@ -15,7 +15,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 # local imports (problem definition)
-from probeye.definition.inference_problem import InferenceProblem
+from probeye.definition.inverse_problem import InverseProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
@@ -74,10 +74,10 @@ class TestProblem(unittest.TestCase):
         #                         Define the Inference Problem                         #
         # ============================================================================ #
 
-        # initialize the inference problem with a useful name; note that the name will
-        # only be stored as an attribute of the InferenceProblem and is not important
+        # initialize the inverse problem with a useful name; note that the name will
+        # only be stored as an attribute of the InverseProblem and is not important
         # for the problem itself; can be useful when dealing with multiple problems
-        problem = InferenceProblem("Max likelihood for linear regression")
+        problem = InverseProblem("Max likelihood for linear regression")
 
         # add all parameters to the problem; the first argument states the parameter's
         # global name (here: 'm', 'b' and 'sigma'); the second argument defines the
