+++ tests/integration_tests/tests_without_correlation/test_two_models.py
@@ -17,7 +17,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 # local imports (problem definition)
-from probeye.definition.inference_problem import InferenceProblem
+from probeye.definition.inverse_problem import InverseProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
@@ -131,8 +131,8 @@ class TestProblem(unittest.TestCase):
         #                         Define the Inference Problem                         #
         # ============================================================================ #
 
-        # initialize the inference problem with a useful name
-        problem = InferenceProblem(
+        # initialize the inverse problem with a useful name
+        problem = InverseProblem(
             "Two models with shared parameter and normal additive error"
         )
