+++ tests/unit_tests/postprocessing/test_sampling.py
@@ -2,7 +2,7 @@
 import unittest
 
 # local imports
-from probeye.definition.inference_problem import InferenceProblem
+from probeye.definition.inverse_problem import InverseProblem
 from probeye.postprocessing.sampling import create_pair_plot
 from probeye.postprocessing.sampling import create_posterior_plot
 from probeye.postprocessing.sampling import create_trace_plot
@@ -11,7 +11,7 @@ from probeye.postprocessing.sampling import create_trace_plot
 class TestProblem(unittest.TestCase):
     def test_not_implemented_pair_plot(self):
         # check for currently not implemented plotting frameworks
-        p = InferenceProblem("TestProblem")
+        p = InverseProblem("TestProblem")
         p.add_parameter("a", "model", prior=("normal", {"loc": 0, "scale": 1}))
         # only one parameter given (no error, just warning)
         create_pair_plot(None, p, plot_with="arviz")
@@ -26,7 +26,7 @@ class TestProblem(unittest.TestCase):
 
     def test_not_implemented_create_posterior_plot(self):
         # check for currently not implemented plotting frameworks
-        p = InferenceProblem("TestProblem")
+        p = InverseProblem("TestProblem")
         with self.assertRaises(NotImplementedError):
             create_posterior_plot(None, p, plot_with="seaborn")
         with self.assertRaises(NotImplementedError):
@@ -36,7 +36,7 @@ class TestProblem(unittest.TestCase):
 
     def test_not_implemented_create_trace_plot(self):
         # check for currently not implemented plotting frameworks
-        p = InferenceProblem("TestProblem")
+        p = InverseProblem("TestProblem")
         with self.assertRaises(NotImplementedError):
             create_trace_plot(None, p, plot_with="seaborn")
         with self.assertRaises(NotImplementedError):
